Add Jenkins test script
jenkins-test can be used to run unit and acceptance tests on heroku.
Usage: `./jenkins-test.py <PR-number>`
It waits for the review app to be ready and then runs unit and acceptance
tests. If exits with error in case of problems or just finishes otherwise.

This is still just in early version, it should be tested and completed
properly.